Address issues from review
