+++ src/api/index.ts
@@ -38,3 +38,4 @@ export { DistributionAPI } from './distribution';
 export { MyDistributionAPI } from './my-distribution';
 export { DistributionType } from './response-types/distribution';
 export { ExecutionEnvironmentAPI } from './execution-environment';
+export { ExecutionEnvironmentType } from './response-types/execution-environment';
