+++ src/api/response-types/execution-environment.ts
@@ -1,0 +1,6 @@
+export class ExecutionEnvironmentType {
+  pulp_created: string;
+  name: string;
+  description: string;
+  latest_version_href: string;
+}
